requirements.txt
asgiref==3.5.2
autopep8==2.0.2
beautifulsoup4==4.11.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
Django==4.1.4
django-crispy-forms==2.0
fonttools==4.38.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mplcursors==0.5.2
murmurhash==1.0.9
networkx==3.0
nltk==3.8.1
numpy==1.24.1
packaging==23.0
pathy==0.10.1
Pillow==9.4.0
platformdirs==3.8.1
plotly==5.15.0
preshed==3.0.8
pycodestyle==2.10.0
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2022.10.31
requests==2.28.2
six==1.16.0
smart-open==6.3.0
soupsieve==2.3.2.post1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.5
tenacity==8.2.2
thinc
tomli==2.0.1
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
wasabi==1.1.1
yapf==0.40.1
zipp==3.15.0

transformers
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
en_use_lg @ https://github.com/MartinoMensio/spacy-universal-sentence-encoder/releases/download/v0.4.6/en_use_lg-0.4.6.tar.gz#en_use_lg-0.4.6
scikit-learn
evaluate
yake
openai
python-dotenv
